refactor: extract shared tag color tokens utility; add missing test case
- Create src/app/utils/tag-color-tokens.ts with tagChipClass, tagDotClass, tagInlineColors
- Update GroupNode.tsx to import tagChipClass/tagDotClass from shared util (removes inline duplicate)
- Update ClassNode.tsx to use tagInlineColors from shared util (removes inline colorMap)
- Add 'preserves existing metadata tags when tags are omitted' test case to helper-group-metadata-sync.test.ts

Agent-Logs-Url: https://github.com/KenSuenobu/objectified-commercial/sessions/777b5f67-52ff-4df7-81a3-72b109387729
