+++ objectified-ui/src/app/components/ade/studio/GroupNode.tsx
@@ -11,6 +11,7 @@ import {
 } from 'lucide-react';
 import * as Popover from '@radix-ui/react-popover';
 import { COLLAPSED_GROUP_FRAME_WIDTH, COLLAPSED_GROUP_FRAME_HEIGHT } from '@/app/utils/canvas-group-collapse';
+import { tagChipClass, tagDotClass } from '@/app/utils/tag-color-tokens';
 import GroupBulkEditDialog, { type GroupBulkEditTagOption } from './GroupBulkEditDialog';
 import {
   Select,
@@ -64,33 +65,6 @@ export const SHADOW_OPTIONS = [
   { name: 'md', label: 'Medium', class: 'shadow-md' },
   { name: 'lg', label: 'Large', class: 'shadow-lg' },
 ];
-
-/** Tailwind classes for project tag chips (matches class-node tag colors). */
-function tagChipClass(color: string): string {
-  const m: Record<string, string> = {
-    default: 'bg-gray-100 dark:bg-gray-700 text-gray-800 dark:text-gray-100 border-gray-200 dark:border-gray-600',
-    primary: 'bg-indigo-100 dark:bg-indigo-900/40 text-indigo-800 dark:text-indigo-200 border-indigo-200 dark:border-indigo-700',
-    secondary: 'bg-purple-100 dark:bg-purple-900/40 text-purple-800 dark:text-purple-200 border-purple-200 dark:border-purple-700',
-    error: 'bg-red-100 dark:bg-red-900/40 text-red-800 dark:text-red-200 border-red-200 dark:border-red-700',
-    warning: 'bg-amber-100 dark:bg-amber-900/40 text-amber-900 dark:text-amber-200 border-amber-200 dark:border-amber-700',
-    info: 'bg-blue-100 dark:bg-blue-900/40 text-blue-800 dark:text-blue-200 border-blue-200 dark:border-blue-700',
-    success: 'bg-emerald-100 dark:bg-emerald-900/40 text-emerald-800 dark:text-emerald-200 border-emerald-200 dark:border-emerald-700',
-  };
-  return m[color] || m.default;
-}
-
-function tagDotClass(color: string): string {
-  const m: Record<string, string> = {
-    default: 'text-gray-500',
-    primary: 'text-indigo-500',
-    secondary: 'text-purple-500',
-    error: 'text-red-500',
-    warning: 'text-amber-500',
-    info: 'text-blue-500',
-    success: 'text-emerald-500',
-  };
-  return m[color] || m.default;
-}
 
 // Predefined group colors with name, bg, border, and text colors
 export const GROUP_COLORS = [
